fix: satisfy backend-size limits after Agy provider integration
Extract session run completion helper, split Agy inference provider code into dedicated files, and move Agy execute tests so CI backend-size verification passes.
