+++ pkg/workers/provider/inference_provider_test.go
@@ -15,7 +15,6 @@ import (
 
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 	"github.com/portpowered/infinite-you/pkg/logging"
-	"github.com/portpowered/infinite-you/pkg/workers/agypty"
 	cursorpkg "github.com/portpowered/infinite-you/pkg/workers/provider/cursor"
 )
 
@@ -995,57 +994,6 @@ func TestScriptWrapProvider_Infer_ExitFailureIncludesExitCodeAndProcessOutput(t 
 	// 		t.Fatalf("expected error %q to contain %q", got, want)
 	// 	}
 	// }
-}
-
-func TestScriptWrapProviderExecuteAgyUsesPTYAdapterPath(t *testing.T) {
-	t.Parallel()
-
-	factoryRoot := t.TempDir()
-	mock := &agyInferenceStubAllocator{result: agypty.SessionResult{ExitCode: 0, CleanedText: "final answer"}}
-	provider := NewScriptWrapProvider(
-		WithAgyFactoryRoot(factoryRoot),
-		WithAgyPTYAllocator(mock),
-	)
-	response, err := provider.Execute(context.Background(), interfaces.RunnerExecutionRequest{
-		Dispatch:         interfaces.WorkDispatch{DispatchID: "dispatch-agy-cli"},
-		ModelProvider:    string(interfaces.ModelProviderAgy),
-		WorkingDirectory: ".",
-		UserMessage:      "plan the goal",
-	})
-	if err != nil {
-		t.Fatalf("Execute() error = %v", err)
-	}
-	if response.Content != "final answer" {
-		t.Fatalf("content = %q, want final answer", response.Content)
-	}
-	if len(mock.sessions) != 1 {
-		t.Fatalf("pty sessions = %d, want 1", len(mock.sessions))
-	}
-	if err := agypty.ValidateArgv(mock.sessions[0].launch.Argv); err != nil {
-		t.Fatalf("ValidateArgv() error = %v", err)
-	}
-}
-
-type agyInferenceStubSession struct {
-	launch agypty.ProcessLaunch
-	result agypty.SessionResult
-}
-
-func (s *agyInferenceStubSession) Run(context.Context) (agypty.SessionResult, error) {
-	return s.result, nil
-}
-
-func (s *agyInferenceStubSession) Close() error { return nil }
-
-type agyInferenceStubAllocator struct {
-	sessions []*agyInferenceStubSession
-	result   agypty.SessionResult
-}
-
-func (a *agyInferenceStubAllocator) Allocate(_ context.Context, launch agypty.ProcessLaunch, _ agypty.SessionConfig) (agypty.PTYSession, error) {
-	session := &agyInferenceStubSession{launch: launch, result: a.result}
-	a.sessions = append(a.sessions, session)
-	return session, nil
 }
 
 // portos:func-length-exception owner=agent-factory reason=legacy-codex-provider-error-table review=2026-07-19 removal=split-command-output-fixtures-and-shared-contract-assertions-before-next-provider-normalization-change
